+++ includes/Request.hpp
@@ -3,90 +3,79 @@
 
 # include <iostream>
 # include <vector>
+# include <map>
 # include "libft.hpp"
+# include "enums.hpp"
+# include "server.hpp"
+# include "config.hpp"
 
 # define NEEDBODY -1
 # define REQUESTCOMPLETE -2
+
+# define AUTHRIZATION "Authorization"
+
+/*
+Request 책임/권한
+- Request message 파싱
+- Request message의 유효성 판단
+- cgi 여부 판단
+- 주요 변경 header -> map< key, value > 로 변경
+*/
+
+class Client;
 
 class Request
 {
 	private:
 		std::string	raw_request;
-
+		
 		std::string	method;
 		std::string	uri;
 		std::string	http_version;
-
-		std::string	accept_charsets;
-		std::string	accept_language;
-		std::string	authorization;
-		std::string	content_length;
-		std::string	content_type;
-		std::string	date;
-		std::string	host;
-		std::string	referer;
-		std::string	transfer_encoding;
-		std::string	user_agent;
-		std::string raw_header;
+		std::map<std::string, std::string> headers;
 		std::string	raw_body;
-
 		std::string temp_body;
-		int	status;
-		int	type;
+		bool	body_receving; //0 이면 헤더 받는 중 , 1 이면 바디 작성.
+		e_body_length_info	body_length_info_type;
+
+		Client * client;
 
 	public:
+		//constructor and destroyer
 		Request(void);
-		Request(const Request& src);
 		virtual ~Request(void){};
-		Request& operator=(const Request& src);
 
-		std::string&	getRawRequest(void);
-		std::string&	getRawBody(void);
-		const std::string&	getRawBody(void) const;
+		//methods
+		void	initRequest(void);
+		e_request_try_make_request_return	tryMakeRequest(void);
+		//setter
+		void				setClient(Client *client);
+
+		//getter
+		std::string&		getRawRequest(void);
 		const std::string&	getMethod(void) const;
 		const std::string&	getUri(void) const;
 		const std::string&	getHttpVersion(void) const;
-		const std::string&	getAcceptCharsets(void) const;
-		const std::string&	getAcceptLanguage(void) const;
-		const std::string&	getAuthorization(void) const;
-		const std::string&	getContentLength(void) const;
-		const std::string&	getContentType(void) const;
-		const std::string&	getDate(void) const;
-		const std::string&	getHost(void) const;
-		const std::string&	getReferer(void) const;
-		const std::string&	getTransferEncoding(void) const;
-		const std::string&	getUserAgent(void) const;
+		std::map<std::string, std::string>&	getHeaders(void) const;
+		const std::string&	getRawBody(void) const;
 
-		void	initRequest(void);
-		bool	tryMakeRequest(void);
+
 	private:
+		int		 	base64_decode(const char * text, char * dst, int numBytes);
+		e_request_try_make_request_return	requestValidCheck(bool isComplete);
+
 		void	makeStartLine(void);
 		void	makeRequestHeader(void);
-		void	makeRequestBody(void);
 
 		void	parseMethod(void);
 		void	parseUri(void);
 		void	parseHttpVersion(void);
 
-		void	parseAcceptCharsets(void);
-		void	parseAcceptLanguage(void);
-		void	parseAuthorization(void);
-		void	parseContentLength(void);
-		void	parseContentType(void);
-		void	parseDate(void);
-		void	parseHost(void);
-		void	parseReferer(void);
-		void	parseTransferEncoding(void);
-		void	parseUserAgent(void);
-
 		bool	bodyCheck(void);
 		bool	isComplete(void);
 
-
-// 테스트용 메소드
-	public:
-		std::string	createRawRequest(void) const;
-		void	bodyPrint(void);
+		bool	isValidAuthHeader(Location &loc);
+		bool	isValidMethod(Location &loc);
 };
 
 #endif
