+++ includes/nginx.hpp
@@ -14,9 +14,19 @@
 # include <fcntl.h>
 # include <sys/stat.h>
 # include <vector>
-# include "parser.hpp"
+# include "config.hpp"
+# include "server.hpp"
+# include "client.hpp"
+# include "resource.hpp"
 
-class Nginx
+/*
+Nginx 
+책임
+- select()를 통해 모든 fd를 관리
+> 생성자/소멸자/
+*/
+
+class Nginx 
 {
 	private :
 		fd_set	reads;
@@ -24,21 +34,27 @@ class Nginx
 		fd_set	errors;
 
 		int		fd_max;
-
-		std::vector<Fdmanager *> fds;
+		std::vector<Fdmanager *> fd_pool;  /// 모든 fd pool (Server, Client, Resource가 담긴다.)
 
 	public	:
+		// 생정자 & 소멸자
 		Nginx();
 		virtual ~Nginx();
 
-		bool	initServers(int queue_size);
-		bool	run(struct timeval	timeout, unsigned int buffer_size);
-
-		void	clear_connected_socket(Fdmanager * fdmanager);
-		void 	insert_pull(Fdmanager *fdmanager);
-		Location &getPerfectLocation(int server_socket_fd, Request		&request);
-
-		std::vector<Fdmanager *>  & getFds();
+		// public method
+		bool	initServers(); // Config 에 적혀있는 정보를 기반으로 Server 인스턴스를 직접 만들어 내서 자신의 fds 에 집어넣는다.
+		bool	run(); 
+		void	deleteFromFdPool(Fdmanager * fdmanager);
+		void 	insertToFdpool(Fdmanager *fdmanager);
+	private :
+		// run()'s
+		bool	isIndexOfReadFdSet(int index, fd_set &reads);
+		bool	isIndexOfWriteFdSet(int index, fd_set &writes);
+		void	doReadServerFd(int i);
+		void	doReadClientFD(int i);
+		void	doReadResourceFD(int i);
+		void	doWriteClientFD(int i);
+		void	doWriteResourceFD(int i);
 		
 };
