+++ srcs/config.cpp
@@ -1,7 +1,5 @@
-#include "../includes/parser.hpp"
+#include "config.hpp"
 
-/////////////////////////////////////////////////////////
-/////////////////// class Config start //////////////////
 Config* Config::instance;
 
 int const Config::decodeMimeBase64[256] = {
@@ -136,8 +134,6 @@ Config::Config()
 	this->status_code["510"] = "Not Extened";
 	this->status_code["511"] = "Network Authentication Required";
 	this->status_code["599"] = "Network Connect Timeout Error";
-
-	this->nginx = NULL;
 }
 
 Config::Config(const Config &src)
@@ -162,18 +158,6 @@ Config *Config::getInstance()
 	if (instance == NULL)
 		instance = new Config();
 	return (instance);
-}
-
-void	Config::setNginx(Nginx *nginx)
-{
-	if (this->nginx == NULL)
-		this->nginx = nginx;
-	return ;
-}
-
-Nginx	*Config::getNginx()
-{
-	return (this->nginx);
 }
 
 bool	Config::returnFalseWithMsg(const char *str)
@@ -359,476 +343,3 @@ void		Config::show()
 		iter->second.show();
 	}
 }
-/////////////////// class Config end ////////////////////
-/////////////////////////////////////////////////////////
-
-
-/////////////////////////////////////////////////////////
-/////////////// class Fdmanager start ///////////////////
-Fdmanager::Fdmanager() : fd(-1), type(BASE)
-{
-
-}
-
-Fdmanager::~Fdmanager()
-{
-
-}
-
-void Fdmanager::setFd(int fd)
-{
-	this->fd = fd;
-	return ;
-}
-
-void	Fdmanager::setType(e_type type)
-{
-	this->type = type;
-	return ;
-}
-
-e_type	Fdmanager::getType() const
-{
-	return (this->type);
-}
-
-int		Fdmanager::getFd() const
-{
-	return (this->fd);
-}
-
-//////////////// class Fdmanager end ////////////////////
-/////////////////////////////////////////////////////////
-
-/////////////////////////////////////////////////////////
-/////////////// class Resource start ////////////////////
-Resource::Resource() : fd_read_to(-1), fd_write_from(-1)
-{
-	this->type = RESOURCE;
-}
-
-Resource::~Resource()
-{
-
-}
-
-void	Resource::setFdReadTo(int fd_read_to)
-{
-	this->fd_read_to = fd_read_to;
-	return ;
-}
-
-void	Resource::setFdWriteFrom(int fd_write_from)
-{
-	this->fd_write_from = fd_write_from;
-	return ;
-}
-
-int		Resource::getFdReadTo()
-{
-	return (this->fd_read_to);
-}
-
-int		Resource::getFdWriteFrom()
-{
-	return (this->fd_write_from);
-}
-
-//////////////// class Resource end /////////////////////
-/////////////////////////////////////////////////////////
-
-/////////////////////////////////////////////////////////
-//////////////////// class Pipe start ///////////////////
-Pipe::Pipe() : pipe_read(-1), pipe_write(-1), write_from_client(-1), read_from_fd(-1)
-{
-	this->type = PIPE;
-}
-
-Pipe::Pipe(int fd) : pipe_read(-1), pipe_write(-1),write_from_client(-1), read_from_fd(-1)
-{
-	this->fd = fd;
-	this->type = PIPE;
-}
-
-Pipe::~Pipe()
-{
-
-}
-
-void	Pipe::setPipeRead(int pipe_read)
-{
-	this->pipe_read = pipe_read;
-	return ;
-}
-
-void	Pipe::setPipeWrite(int pipe_write)
-{
-	this->pipe_write = pipe_write;
-	return ;
-}
-
-void	Pipe::setReadFromFd(int read_from_fd)
-{
-	this->read_from_fd = read_from_fd;
-	return ;
-}
-
-void	Pipe::setWriteFromClient(int write_from_client)
-{
-	this->write_from_client = write_from_client;
-	return ;
-}	
-
-int		Pipe::getPipeRead()
-{
-	return (this->pipe_read);
-}
-
-int		Pipe::getPipeWrite()
-{
-	return (this->pipe_write);
-}
-
-int		Pipe::getReadFromFd()
-{
-	return (this->read_from_fd);
-}
-
-int		Pipe::getWriteFromClient()
-{
-	return (this->write_from_client);
-}
-//////////////////// class Pipe end /////////////////////
-/////////////////////////////////////////////////////////
-
-/////////////////////////////////////////////////////////
-////////////////// class Client start ///////////////////
-Client::Client()
-{
-	this->server_socket_fd = -1;
-	this->fd = -1;
-	this->status = REQUEST_RECEIVING;
-	this->type = CLIENT;
-}
-
-Client::Client(int server_socket_fd, int fd) : server_socket_fd(server_socket_fd)
-{
-	this->fd = fd;
-	this->status = REQUEST_RECEIVING;
-	this->type = CLIENT;
-	this->server_socket_fd = -1;
-}
-
-Client::~Client()
-{
-
-}
-
-void	Client::setLastRequestMs(unsigned long last_request_ms)
-{
-	this->last_request_ms = last_request_ms;
-}
-
-void	Client::setStatus(t_status status)
-{
-	this->status = status;
-	return ;
-}
-
-void	Client::setServerSocketFd(int server_socket_fd)
-{
-	this->server_socket_fd = server_socket_fd;
-	return ;
-}
-
-t_status	Client::getStatus()
-{
-	return (this->status);
-}
-
-int			Client::getServerSocketFd()
-{
-	return (this->server_socket_fd);
-}
-
-Request		&Client::getRequest()
-{
-	return (this->request);
-}
-
-Response	&Client::getResponse()
-{
-	return (this->response);
-}
-
-unsigned long	Client::getLastRequestMs()
-{
-	return (this->last_request_ms);
-}
-/////////////////// class Client end ////////////////////
-/////////////////////////////////////////////////////////
-
-/////////////////////////////////////////////////////////
-////////////////// class Server start ///////////////////
-Server::Server() : port(-1)
-{
-	this->type = SERVER;
-}
-
-Server::Server(const Server& src)
-{
-	this->type = SERVER;
-	this->ip = src.ip;
-	this->port	=	src.port;
-	this->server_name	=	src.server_name;
-	this->fd		=	src.fd;
-	this->locations.insert(src.locations.begin(), src.locations.end());
-}
-
-Server &Server::operator=(const Server &src)
-{
-	this->type = SERVER;
-	this->ip = src.ip;
-	this->port	=	src.port;
-	this->server_name	=	src.server_name;
-	this->fd		=	src.fd;
-	this->locations.clear();
-	this->locations.insert(src.locations.begin(), src.locations.end());
-
-	return (*this);
-}
-
-Server::~Server()
-{
-	return ;
-}
-
-void	Server::setPort(unsigned short port)
-{
-	this->port = port;
-	return ;
-}
-
-void	Server::setIP(const std::string &ip)
-{
-	this->ip = ip;
-	return ;
-}
-
-void	Server::setServerName(const std::string &server_name)
-{
-	this->server_name = server_name;
-	return ;
-}
-
-const std::string &Server::getIP() const
-{
-	return (this->ip);
-}
-
-const std::string &Server::getServerName() const
-{
-	return (this->server_name);
-}
-
-unsigned short		Server::getPort() const
-{
-	return (this->port);
-}
-
-std::map<std::string, Location> &Server::getLocations()
-{
-	return (this->locations);
-}
-
-//for test
-void		Server::show()
-{
-	std::cout << "ip	:	" << this->ip << std::endl;
-	std::cout << "port	:	" << this->port << std::endl;
-	std::cout << "server_name	:	" << this->server_name << std::endl;
-	std::cout << "============= location start =============" << std::endl;
-	for (std::map<std::string, Location>::iterator iter = locations.begin(); iter != locations.end(); iter++)
-	{
-		std::cout << "=== Location Key : " << iter->first << " ===" << std::endl;
-		iter->second.show();
-	}
-	std::cout << "============= location end ===============" << std::endl;
-}
-
-////////////////// class Server end /////////////////////
-/////////////////////////////////////////////////////////
-
-/////////////////////////////////////////////////////////
-///////////////// class Location start //////////////////
-Location::Location() : request_max_body_size(-1), redirect_return(-1)
-{
-	
-}
-
-Location::Location(const Location &src)
-{
-	this->root	=	src.root;
-	this->index.assign(src.index.begin(), src.index.end());
-	this->allow_methods.assign(src.allow_methods.begin(), src.allow_methods.end());
-	this->request_max_body_size = src.request_max_body_size;
-	this->error_pages.insert(src.error_pages.begin(), src.error_pages.end());
-	this->upload_path = src.upload_path;
-	this->auto_index = src.auto_index;
-	this->cgi_infos.insert(src.cgi_infos.begin(), src.cgi_infos.end());
-	this->auth_key = src.auth_key;
-	this->redirect_addr = src.redirect_addr;
-	this->redirect_return = src.redirect_return;
-	this->uri_key = src.uri_key;
-}
-
-Location &Location::operator=(const Location &src)
-{
-	this->root	=	src.root;
-	this->index.assign(src.index.begin(), src.index.end());
-	this->allow_methods.assign(src.allow_methods.begin(), src.allow_methods.end());
-	this->request_max_body_size = src.request_max_body_size;
-	this->error_pages.insert(src.error_pages.begin(), src.error_pages.end());
-	this->upload_path = src.upload_path;
-	this->auto_index = src.auto_index;
-	this->cgi_infos.insert(src.cgi_infos.begin(), src.cgi_infos.end());
-	this->auth_key = src.auth_key;
-	this->redirect_addr = src.redirect_addr;
-	this->redirect_return = src.redirect_return;
-	this->uri_key = src.uri_key;
-	return (*this);
-}
-
-void		Location::setRoot(const std::string& root)
-{
-	this->root = root;
-	return ;
-}
-
-void		Location::setRequestMaxBodySize(int request_max_body_size)
-{
-	this->request_max_body_size = request_max_body_size;
-	return ;
-}
-void		Location::setUploadPath(const std::string &upload_path)
-{
-	this->upload_path = upload_path;
-	return ;
-}
-
-void		Location::setAutoIndex(bool auto_index)
-{
-	this->auto_index = auto_index;
-	return ;
-}
-
-void		Location::setAuthKey(const std::string &auth_key)
-{
-	this->auth_key = auth_key;
-	return ;
-}
-
-void		Location::setRedirectReturn(int redirect_return)
-{
-	this->redirect_return = redirect_return;
-	return ;
-}
-
-void		Location::setRedirectAddr(const std::string &redirect_addr)
-{
-	this->redirect_addr = redirect_addr;
-	return ;
-}
-
-void		Location::setUriKey(const std::string &uri_key)
-{
-	this->uri_key = uri_key;
-	return ;
-}
-
-const std::string &Location::getUriKey()
-{
-	return (this->uri_key);
-}
-
-const std::string &Location::getRoot()
-{
-	return (this->root);
-}
-
-std::list<std::string> &Location::getIndex()
-{
-	return (this->index);
-}
-
-std::list<std::string> &Location::getAllowMethods()
-{
-	return (this->allow_methods);
-}
-
-int Location::getRequestMaxBodySize()
-{
-	return (this->request_max_body_size);
-}
-
-std::map<int, std::string> &Location::getErrorPages()
-{
-	return (this->error_pages);
-}
-
-const std::string &Location::getUploadPath()
-{
-	return (this->upload_path);
-}
-
-bool	Location::getAutoIndex()
-{
-	return (this->auto_index);
-}
-
-std::map<std::string, std::string>& Location::getCgiInfos()
-{
-	return (this->cgi_infos);
-}
-
-const std::string &Location::getAuthKey()
-{
-	return (this->auth_key);
-}
-
-int		Location::getRedirectReturn()
-{
-	return (this->redirect_return);
-}
-
-const std::string &Location::getRedirectAddr()
-{
-	return (this->redirect_addr);
-}
-
-//for test
-void	Location::show()
-{
-	std::cout << "root	:	" << this->root << std::endl;
-	std::cout << "rqmbs	:	" << this->request_max_body_size << std::endl;
-	std::cout << "upload_path	:	" << this->upload_path << std::endl;
-	std::cout << "auto_index	:	" << this->auto_index << std::endl;
-	std::cout << "auth_key	:	" << this->auth_key << std::endl;
-	std::cout << "index	: ";
-	for (std::list<std::string>::iterator iter = this->index.begin(); iter != this->index.end(); iter++)
-		std::cout << *iter << " ";
-	std::cout << std::endl;
-	std::cout << "allow_methods	: ";
-	for (std::list<std::string>::iterator iter = this->allow_methods.begin(); iter != this->allow_methods.end(); iter++)
-		std::cout << *iter << " ";
-	std::cout << std::endl;
-	std::cout << "error_pages	: " << std::endl;
-	for (std::map<int, std::string>::iterator iter = this->error_pages.begin(); iter != this->error_pages.end(); iter++)
-		std::cout << iter->first << " | " << iter->second << std::endl;
-	std::cout << "cgi_infos	: " << std::endl;
-	for (std::map<std::string, std::string>::iterator iter = this->cgi_infos.begin(); iter != this->cgi_infos.end(); iter++)
-		std::cout << iter->first << " | " << iter->second << std::endl;
-}
-
-///////////////// class Location end ////////////////////
-/////////////////////////////////////////////////////////
