+++ srcs/main.cpp
@@ -1,6 +1,6 @@
-#include "../libft_cpp/libft.hpp"
-#include "../includes/parser.hpp"
-#include "../includes/nginx.hpp"
+#include "libft.hpp"
+#include "config.hpp"
+#include "nginx.hpp"
 
 int	main(int ac, char **av)
 {
@@ -12,16 +12,14 @@ int	main(int ac, char **av)
 	//return 0;
 
 	Nginx nginx;
-	struct timeval		timeout;
+//	struct timeval		timeout;
 
-	Config::getInstance()->setNginx(&nginx);
-
-	timeout.tv_sec = 5; // last request time out 5000ms
-	timeout.tv_usec = 0;
+	// timeout.tv_sec = 5; // last request time out 5000ms
+	// timeout.tv_usec = 0;
 	try
 	{
-		nginx.initServers(5);
-		nginx.run(timeout, 65536);
+		nginx.initServers();
+		nginx.run();
 	}
 	catch(const char *e)
 	{
